Reorg in drink-debug
